Improve error handling
This commit improves the error handling in the case that
write errors are detected. If a reopen() function is available
it is called to reopen stale file handle.
Tests with failfs however have identified a problem that
is not resolved with the current code version: If a file system
is remounted it might happen that watchman reopens the file during
the time where no filesystem is mounted thus writing to the
underlying filesystem and potentially blocking the mounting of
the wanted filesystem.