Add class Named_Clingy_File
This commit adds the new class Named_Clingy_File which
inherits from Named_File. In contrast to a Named_File
instance a clingy file will only allow reopen() to be called
if the filesystem to which the file is attached (may [usually it
is] or may not be the one on which the file resides) is
actually mounted. If the file resides on a clustered filesystem
that makes sure that the file is not accidentally reopened on
the local filesystem/disk during a remount of the parallel
filesystem.